Add files via upload
Open Excel and navigate to the Developer tab. If the Developer tab is not visible, you will need to enable it in the Excel options.
In the Developer tab, select Import and choose the file you want to import.
After importing the file, go back to the Developer tab and select Visual Basic. This will open the Visual Basic for Applications (VBA) editor.
In the VBA editor, look at the Project Explorer on the left side. Right-click on the project into which you want to insert the file, and select Insert.
Navigate to the file you want to insert and select it.
Once the file is inserted, go to the Macros dialog box (you can access this from the Developer tab in Excel or by pressing Alt + F8).
In the Macros dialog box, find and select the CBI_BANK_STAT() macro, then click Run.
Please ensure that you have the necessary permissions and that macros are enabled in your Excel settings. Always be cautious when running macros from files you received from others, as they can potentially contain harmful code. If you’re unsure, please consult with your IT department or a trusted advisor.